Remove autocorrection logic
